Rename desmume_utils module
