+++ tests/desmume/emulator_utils.py
